Shiny module for selecting data

Shiny module for selecting data

dataSelectUI(id)

dataSelectServer(id, data = reactive(NULL), hide = FALSE)

Arguments

id
unique identifier for the module to prevent namespace clashes when making multiple calls to this shiny module.

data
an array wrapped in reactive() containing the data to be filtered.

hide
logical indicating whether the data selection user interface should be hidden from the user, set to FALSE by default.

Value

a list of reactive objects containing the filtered data and indices for selected columns.

Examples

if (interactive()) {
  library(shiny)
  library(rhandsontable)
  library(shinyjs)

  ui <- fluidPage(
    useShinyjs(),
    dataInputUI("input1"),
    dataSelectUI("select1"),
    dataOutputUI("output1"),
    rHandsontableOutput("data1")
  )

  server <- function(input,
                     output,
                     session) {
    data_input <- dataInputServer("input1")

    data_select <- dataSelectServer("select1",
      data = data_input
    )

    output$data1 <- renderRHandsontable({
      if (!is.null(data_select())) {
        rhandsontable(data_select())
      }
    })

    dataOutputServer("output1",
      data = data_select
    )
  }

  shinyApp(ui, server)
}