An interactive editor for viewing, entering & editing data

data_edit is a shiny application built on rhandsontable that is designed to make it easy to interactively view, enter or edit data without any coding. data_edit is also a wrapper for any reading or writing function to make it easy to interactively update data saved to file.

data_edit(
  x,
  col_bind = NULL,
  col_edit = TRUE,
  col_options = NULL,
  col_stretch = FALSE,
  col_factor = FALSE,
  col_names = TRUE,
  row_bind = NULL,
  row_edit = TRUE,
  save_as = NULL,
  title = NULL,
  logo = NULL,
  logo_size = 100,
  viewer = TRUE,
  theme = "yeti",
  quiet = FALSE,
  read_fun = "read.csv",
  read_args = NULL,
  write_fun = "write.csv",
  write_args = NULL,
  ...
)

Arguments

x
a matrix, data.frame, data.table or the name of a csv file to edit.Tibble are also supported but will be coerced to data.frames. An empty table can be created by specifying the dimensions in a vector of the form c(nrow, ncol).

col_bind
additional columns to add to the data prior to loading into editor, can be either an array containing the new data, a vector containing the new column names for empty columns or a named list containing a vector for each new column.

col_edit
logical indicating whether columns can be added or removed, set to TRUE by default.

col_options
named list containing the options for columns that use dropdown menus or checkboxes.

col_stretch
logical indicating whether columns should be stretched to fill the full width of the display, set to FALSE by default.

col_factor
logical indicating whether character columns should be converted to factors prior to returning the edited data, set to FALSE by default.

col_names
logical indicating whether column names can be edited or a vector of column names that cannot be edited, set to TRUE by default to allow editing of column names.

row_bind
additional rows to add to the data prior to loading into editor, can be either an array containing the new data, a vector containing the new row names for empty rows or a named list containing a vector for each new column.

row_edit
logical indicating whether rows can be added or removed, set to TRUE by default.

save_as
name of a csv file to which the edited data should be saved.

title
optional title to include above the data editor.

logo
optional package logo to include in title above the data editor, must be supplied as path to logo png.

logo_size
width of the logo in pixels, set to 100 pixels by default.

viewer
logical indicating whether the data editor should be invoked in the RStudio viewer pane, set to TRUE by default.

theme
valid shinytheme name, set to "yeti" by default.

quiet
logical indicating whether messages should be suppressed, set to FALSE by default.

read_fun
name of the function to use to read in the data when x is the name of a file, set to read.csv by default.

read_args
a named list of additional arguments to pass to read_fun.

write_fun
name of the function to use to write the edited version of x to a file, set to write.csv by default. Only requirement is that the first argument accepts the edited data and the second argument accepts the file name supplied to save_as.

write_args
a named list of additional arguments to pass to write_fun.

...
additional arguments passed to both read_fun and write_fun, such as sep. Arguments that are shared by read_fun and write_fun can be supplied separately to these functions using the read_args and write_args arguments. For example, this becomes particularly important when specifying row.names arguments.

Value

edited matrix-like object.

Examples

if (interactive()) {
  # Edit matrix & save to csv
  data_edit(mtcars,
    save_as = "mtcars-update.csv"
  )

  # Edit csv file
  data_edit("mtcars-update.csv")
}